separated the overlay and calculation into a different file and imported it as a component on the map
